Move common models and functions into models/base.py
